grader: allow percentage output to be floating point

+++ judgels-backends/judgels-grader-engines/src/main/java/judgels/gabriel/helpers/TestCaseVerdictParser.java
@@ -15,7 +15,7 @@ public class TestCaseVerdictParser {
 
         Verdict verdict;
         Optional<Double> points = Optional.empty();
-        Optional<Integer> percentage = Optional.empty();
+        Optional<Double> percentage = Optional.empty();
         Optional<String> feedback = Optional.empty();
 
         switch (lines[0]) {
@@ -44,7 +44,7 @@ public class TestCaseVerdictParser {
                 if (!result.isEmpty() && result.charAt(result.length() - 1) == '%') {
                     String percentageString = result.substring(0, result.length() - 1);
                     try {
-                        percentage = Optional.of(Integer.parseInt(percentageString));
+                        percentage = Optional.of(Double.parseDouble(percentageString));
                     } catch (NumberFormatException e) {
                         throw new ScoringException("Invalid <percentage> for OK: " + result);
                     }
